create server dir
