--- packages/frontend/rewrite/src/dotenv.ts
@@ -1,4 +1,0 @@
-import { join } from 'node:path'
-import dotenv from 'dotenv'
-
-dotenv.config({ path: join(process.cwd(), '.env.local') })
